Initial migration to nithinmurali/pygsheets ...
Speed of a row insert has gone from a minute and a half with gspread to under a second with pygsheets. 

Need to find out how to silence the call to pygsheets.authorize, because having the service account printed out every time you run the program just doesn't look very good.